Added return types to functions. Renamed memory to vectors
