+++ packages/agent-utils/src/index.ts
@@ -3,4 +3,4 @@ export * from "./sys";
 export * from "./scripts";
 export * from "./llm";
 export * from "./chunking";
-export * from "./memory";
+export * from "./vectors";
