+++ packages/agent-utils/src/vectors/LocalDocumentStore.ts
