+++ packages/agent-utils/src/vectors/OpenAIEmbeddingApi.ts
