+++ packages/agent-utils/src/vectors/utils.ts
